more physics updates
rotating bodies can move player, collisisons are fine tunes, list of rules for
builing meshes are developed to comply
with the physics calculations
