+++ Unity_3D_Third_Person_Unity_Game/Assets/Scripts/CharacterControllerTesting.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b2907e0d4abb5c54a9786f47c57b4143
+guid: 150daf1e782711142a885bc89160b045
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
