--- Unity_3D_Third_Person_Unity_Game/Assets/Scripts/CharacterTestingScripts/CharacterControllerTesting.cs
@@ -1,980 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class CharacterControllerTesting : MonoBehaviour
-{
-
-	Transform cameraT;
-    public float distToGround;
-    public bool isGrounded = true;
-    public Vector3 slopeVector = new Vector3(0, 0, 0);
-    public float forwardHeightTop = 1.0f;
-    public float forwardHeightMiddle = 1.0f;
-    public float forwardHeightBottom = 1.0f;
-    public float downwardThickness = 0.1f; // the difference between the ground and the position of the character
-    public float downwardHeight = 2.0f;
-    // might need two height values, center to top and center to bottom
-    public float forwardThickness = 0.1f; // this need to be have the diameter of the total player
-
-    // Bit shift the index of the layer (8) to get a bit mask
-    public int layerMask = 1 << 8;
-    // This would cast rays only against colliders in layer 8.
-
-    private Animator anim;
-	public float animMoveSpeed = 0.0f;
-	public float animDeltaSpeed = 0.01f;
-	public float animIdleSpeed = 0.0f;
-	public float animWalkSpeed = 0.6f;
-	public float animRunSpeed = 1.0f;
-
-	public CharacterState characterState = CharacterState.STATIONARY;
-
-	public float walkSpeed = 0.1f;
-	public float runSpeed = 0.2f;
-
-    // the current gravity value is contained in the universal vector, gravity increment is added to its y component
-    public float gravityIncrement = -0.0981f; // this is the amount each frame the gravity value changes
-	public float maxFallVelocity = -0.1f; // the maximum value to displace the character
-
-	public float groundDistanceCheck = 0.1f;
-
-	public float turnSmoothTime = 0.0f;
-	float turnSmoothVelocity;
-
-	public float speedSmoothTime = 0.1f;
-	float speedSmoothVelocity;
-	float currentSpeed;
-
-	Vector3 PrevPos;
-	Vector3 NewPos;
-	Vector3 ObjVelocity;
-
-
-
-	public float initialJumpVelocity = 1.0f;
-	public float initialHighJumpVelocity = 2.0f;
-	public Vector2 initialLongJumpVelocity = new Vector2(1.0f, 1.0f); // length and height
-    
-    
-    
-
-	// input variables
-	Vector2 input;
-	float jump;
-	float specialJump;
-	float groundPound;
-
-	// secondary input variables
-	public bool running;
-
-
-
-
-	// this is the vector by which the player moves, all inputs and movements are added to this vector and then
-	// this vector is applied to the player each update, so to jump you apply a single up value and on each frame
-	// that value will be used to move the player up, and gravity will reduce it
-	// when grounded the y component will be set to zero
-	public Vector3 universalMovementVector = new Vector3(0, 0, 0);
-	public float slopeSlipSpeed = 0.1f;
-	public bool InputAccepted = true;
-	public Vector3 extraRotationTranslationVector = Vector3.zero;
-	public float groundPoundTimer = 0.0f;
-	public float groundPoundTimeLimit = 4.0f;
-
-    // Start is called before the first frame update
-    void Start(){
-        // But instead we want to collide against everything except layer 8. The ~ operator does this, it inverts a bitmask.
-        layerMask = ~layerMask;
-
-        PrevPos = transform.position;
-		NewPos = transform.position;
-
-		cameraT = Camera.main.transform;
-
-		anim = gameObject.GetComponentInChildren<Animator>();
-		anim.SetFloat("speed", animIdleSpeed);
-	}
-
-	// Update is called once per frame
-	void FixedUpdate(){
-		// first we check for "no input status"
-		// then collect the inputs from the player
-		// enter the function call from current character state
-		// and then change the state respsectively, followed by
-		// acting on the input values
-		if(!InputAccepted){
-			// act like no inputs being made
-			input = new Vector2(0, 0);
-			jump = 0;
-			specialJump = 0;
-			groundPound = 0;
-		}
-		else{
-			// inputs are allowed
-			input = new Vector2 (Input.GetAxisRaw("Horizontal"), Input.GetAxisRaw("Vertical"));
-			input = input.normalized;
-			jump = Input.GetAxisRaw("Jump");
-			specialJump = Input.GetAxisRaw("Special Jump");
-			groundPound = Input.GetAxisRaw("Ground Pound");
-		}
-		// inputs are collected
-
-		if(characterState == CharacterState.STATIONARY){
-			StationaryState();
-		}
-		else if(characterState == CharacterState.RUNNING){
-			RunningState();
-		}
-		else if(characterState == CharacterState.JUMPING_LOW){
-			LowJumpingState();
-		}
-		else if(characterState == CharacterState.JUMPING_LONG){
-			LongJumpingState();
-		}
-		else if(characterState == CharacterState.JUMPING_HIGH){
-			HighJumpingState();
-		}
-		else if(characterState == CharacterState.FALLING){
-			FallingState();
-		}
-		else if(characterState == CharacterState.FALLING_NO_CONTROL){
-
-		}
-		else if(characterState == CharacterState.SLIPPING_NO_CONTROL){
-			SlippingState();
-		}
-		else if(characterState == CharacterState.GROUND_POUND){
-			GroundPoundState();
-		}
-        // no matter what state we are in we must check the collisions with the ray cast
-        // thus we do the collision testing after the state function
-        // each raycast will set values of their own (downward raycast will set isgrounded if it detects a ground)
-        // thus checking for isgrounded is done by variable rather than function
-        // the downward raycast must check for ground and a rotating body, if rotating body is found then
-        // set the addtional translation vector
-        // also we must break down movement into parts to fit with the raycasts
-        // that means the translations must be calulated in each state but not executed
-        // make sure the translation parts are smaller than the smallest raycast used and then raycast for each movement
-        // or we can break down the movement vector into 3 parts x, y, z and move each one and raycast based of the values
-        // so raycast in the x at the maximum value allowed to move forward and then move the player the smaller of the two values
-        // in the x (desired movement or the limited movement), the repeat for z and then y
-        // also check for rotation translation during the raycast, each state only added their own translation, stationary
-        // doesn't do anything so it can't check for rotating bodies
-        // also dont let any state determine movement using time.deltaTime, we will account for that at the end
-        // also cap off the gravity velocity
-        // also DO NOT USE DELTA TIME, only use motion as frame by frame Application.targetFrameRate = 30
-        // it is important that we check the y raycast last as if we are moving down and in a direction (slope) we
-        // want to move down the slope and should do x and z first
-        // since we are breaking down the universalMovementVector into its x, y, z components we can limit the raycasts being used
-        // first assign default values for x, y, z and for each raycast we begin to set limits for their values
-        // once the limits are determined we need to include an offset for the player size, this is done to all 6 values
-        // after all the raycasts are done and before the movement is executed
-        // say the forward is initial limited by 10.0f and the player is 1.0f units radius, if a wall is found
-        // that is 9.0f units infront, then the limit is set to 9.0f and then set to 8.0f for the thickness of the player
-        // thus the player only moves 8.0f units, this is because the wall is 9.0f away from the player center but 8.0f away
-        // from the player's front radius, we want the front of the player right on the wall
-        // then after all limits are determined we move the player in the x, y, z by smaller of the two values (limits and vector)
-
-        bool xzModified = false;
-        bool yModified = false;
-
-        Vector3 XZ = new Vector3(universalMovementVector.x, 0, universalMovementVector.z);
-        Vector3 Y = new Vector3(0, universalMovementVector.y, 0);
-
-
-        RaycastHit hitDownward;
-        // remember the distance is based off the universalMovementVector
-        // we only want the x and z components of the universalMovementVector, x and z limits are identical
-        // can use either for distance, and then set the temps based of which is smaller the current temps or the
-        // new calculated temps
-        // we want the player's thickness included in the raycast
-        //Debug.Log(Y.y + " " + downwardThickness);
-        float trans = -(Y.y - downwardThickness) + downwardHeight;
-        Debug.Log(trans);
-        if (Physics.Raycast(transform.position + new Vector3(0, downwardHeight, 0), -transform.up, out hitDownward, trans, layerMask))
-        {
-            // we found something but the normal.y might be >= 0.09
-            //Debug.Log(hitDownward.normal);
-            if (hitDownward.normal.y > 0.09f)
-            {
-                // a floor is detected
-                Debug.Log("floor is hit");
-                float translator = hitDownward.point.y + downwardThickness;
-                //transform.position = new Vector3(translator.x, transform.position.y, translator.z);
-                transform.position = new Vector3(transform.position.x, translator, transform.position.z);
-                yModified = true;
-                isGrounded = true;
-            }
-            else
-            {
-                isGrounded = false;
-            }
-        }
-        else
-        {
-            Debug.Log("not found");
-            isGrounded = false;
-        }
-
-        //Physics.Raycast(transform.position, transform.forward, out hit, 1.0f, layerMask
-
-
-
-
-
-
-        // these three ray casts must be done at the varying heights of the character
-        RaycastHit hitForwardTop;
-        // remember the distance is based off the universalMovementVector
-        // we only want the x and z components of the universalMovementVector, x and z limits are identical
-        // can use either for distance, and then set the temps based of which is smaller the current temps or the
-        // new calculated temps
-        // we want the player's thickness included in the raycast
-        if (Physics.Raycast(transform.position + new Vector3(0, forwardHeightTop, 0), XZ, out hitForwardTop, XZ.magnitude + forwardThickness, layerMask))
-        {
-            // we found something but the normal.y might be >= 0.09
-            if (hitForwardTop.normal.y < -0.09f)
-            {
-                // a ceiling is detected do nothing, that is for the hitUpward
-            }
-            else if (hitForwardTop.normal.y > 0.09f)
-            {
-                // a floor is detected do nothing, that is for the hitDownward
-            }
-            else
-            {
-                Debug.Log("wall is hit");
-                Vector3 translator = hitForwardTop.point + hitForwardTop.normal * forwardThickness;
-                transform.position = new Vector3(translator.x, transform.position.y, translator.z);
-                xzModified = true;
-            }
-        }
-
-
-
-
-
-        
-        RaycastHit hitForwardMiddle;
-        // remember the distance is based off the universalMovementVector
-        // we only want the x and z components of the universalMovementVector, x and z limits are identical
-        // can use either for distance, and then set the temps based of which is smaller the current temps or the
-        // new calculated temps
-        // we want the player's thickness included in the raycast
-        if (Physics.Raycast(transform.position + new Vector3(0, forwardHeightMiddle, 0), XZ, out hitForwardMiddle, XZ.magnitude + forwardThickness, layerMask))
-        {
-            // we found something but the normal.y might be >= 0.09
-            if (hitForwardMiddle.normal.y <= -0.09f)
-            {
-                // a ceiling is detected do nothing, that is for the hitUpward
-            }
-            else if (hitForwardMiddle.normal.y >= 0.09f)
-            {
-                // a floor is detected do nothing, that is for the hitDownward
-            }
-            else
-            {
-                Debug.Log("wall is hit");
-                Vector3 translator = hitForwardMiddle.point + hitForwardMiddle.normal * forwardThickness;
-                transform.position = new Vector3(translator.x, transform.position.y, translator.z);
-                xzModified = true;
-            }
-        }
-
-        RaycastHit hitForwardBottom;
-        // remember the distance is based off the universalMovementVector
-        // we only want the x and z components of the universalMovementVector, x and z limits are identical
-        // can use either for distance, and then set the temps based of which is smaller the current temps or the
-        // new calculated temps
-        // we want the player's thickness included in the raycast
-        if (Physics.Raycast(transform.position + new Vector3(0, forwardHeightBottom, 0), XZ, out hitForwardBottom, XZ.magnitude + forwardThickness, layerMask))
-        {
-            // we found something but the normal.y might be >= 0.09
-            if (hitForwardBottom.normal.y <= -0.09f)
-            {
-                // a ceiling is detected do nothing, that is for the hitUpward
-            }
-            else if (hitForwardBottom.normal.y >= 0.09f)
-            {
-                // a floor is detected do nothing, that is for the hitDownward
-            }
-            else
-            {
-                Debug.Log("wall is hit");
-                Vector3 translator = hitForwardBottom.point + hitForwardBottom.normal * forwardThickness;
-                transform.position = new Vector3(translator.x, transform.position.y, translator.z);
-                xzModified = true;
-            }
-        }
-
-        MoveCharacter(universalMovementVector, xzModified, yModified);
-
-
-        //Debug.DrawRay(transform.position + new Vector3(0, heightTop, 0), transform.forward * 10, Color.yellow);
-        //Debug.DrawRay(transform.position + new Vector3(0, heightBottom, 0), -transform.up * 10, Color.yellow);
-        //if (Physics.Raycast(transform.position + new Vector3(0, downwardHeight, 0), Y, out hitDownward, -(Y.y - downwardThickness) + downwardHeight + 1.0f, layerMask))
-        //Physics.Raycast(transform.position, Y, out hitDownward, Y.y - downwardThickness, layerMask
-        // a ceiling is defined by any surface with normal.y < -0.09 (5 degrees)
-        // a wall is defined by any surface with -0.09 <= normal.y <= 0.09
-        // a slippery floor is defined by any surface with 0.09 < normal.y < 0.5 (30 degrees = 60 degree slope)
-        // a floor is defined by any surface with 0.5 <= normal.y
-
-
-
-    }
-
-    
-    void MoveCharacter(Vector3 movement, bool xz, bool y)
-    {
-        // do not use Time.deltaTime
-        if (xz)
-        {
-            movement.x = 0;
-            movement.z = 0;
-        }
-        if (y)
-        {
-            movement.y = 0;
-        }
-        transform.Translate(movement, Space.World);
-    }
-
-	void StationaryState(){
-        universalMovementVector = Vector3.zero;
-		if(!isGrounded){
-			// we are in the stationary state but the floor is not present, (got pushed off the edge?)
-			// we swap the state and skip the input
-			characterState = CharacterState.FALLING;
-		}
-		else{
-			if (input != Vector2.zero) {
-				// we have input for running
-				// must exit the stationary state
-				characterState = CharacterState.RUNNING;
-				// notice we can swap into and then out of running state if horizontal and jump
-				// input, this means the character will jump rather than start running from
-				// stationary pose, which means the character can't do a long jump from stationary
-			}
-
-			//CalculateExtraRotationTranslationVector(); // we will check this during the downward ray cast
-			//if(extraRotationTranslationVector != Vector3.zero){
-                // shift character by what the rotation means in translation
-                // universalMovementVector += extraRotationTranslationVector;
-                // we do not move the player in any state, only at the end of the fixed update function
-			//}
-			// now check if a jump has been called
-			// notice how we go to a jump state rather than running state if both running and jumping
-			if(jump > 0 && isGrounded){
-				// later come up with a keyboard click tester (don't let holding down button work as jump input)
-				// now determine what kind of jump
-				if(specialJump > 0){
-					// since we are in the stationary state we will do a high jump
-					HighJump();
-				}
-				else{
-					RegularJump();
-				}
-			}
-			// if no input was registered we remain in the stationary state, and animation will reflect that
-
-			if(animMoveSpeed <= animIdleSpeed){
-				animMoveSpeed = animIdleSpeed;
-			}
-			else{
-				animMoveSpeed -= animDeltaSpeed;
-			}
-
-			anim.SetFloat("speed", animMoveSpeed);
-		}
-	}
-
-	void RunningState(){
-
-		// running covers any horizontal movement, so check speed to do a long jump
-		if(!isGrounded){
-			// we are in the running state but no ground is present, we swap the state
-			// and skip the input
-			characterState = CharacterState.FALLING;
-		}
-		else{
-
-			if (input != Vector2.zero) {
-				// all this is to rotate the character in the desired running direction
-				float targetRotation = Mathf.Atan2 (input.x, input.y) * Mathf.Rad2Deg + cameraT.eulerAngles.y;
-                transform.eulerAngles = Vector3.up * targetRotation;// Mathf.SmoothDampAngle(transform.eulerAngles.y, targetRotation, ref turnSmoothVelocity, turnSmoothTime);
-			}
-			else{
-				// no input so we move back to stationary
-				characterState = CharacterState.STATIONARY;
-			}
-
-			// determine the speed at which to move the player
-			float targetSpeed = ((running || Input.GetKey(KeyCode.E)) ? runSpeed : walkSpeed) * input.magnitude;
-			currentSpeed = Mathf.SmoothDamp (currentSpeed, targetSpeed, ref speedSmoothVelocity, speedSmoothTime);
-
-			// now apply gravity, gravity is m/s^2 so multiply by Time.deltaTime twice (only during a jump or fall)
-			universalMovementVector = transform.forward * currentSpeed;
-			//CalculateExtraRotationTranslationVector();
-
-			//MoveCharacter(universalMovementVector * Time.deltaTime + extraRotationTranslationVector);
-
-			if(jump > 0){
-				// later come up with a keyboard click tester (don't let holding down button work as jump input)
-				// now determine what kind of jump
-				if(specialJump > 0){
-					if(running || Input.GetKey(KeyCode.E)){
-						// since we are in the running state we will do a high jump
-						LongJump();
-					}
-					else{
-						HighJump();
-					}
-				}
-				else{
-					RegularJump();
-				}
-			}
-
-			if(running || Input.GetKey(KeyCode.E)){
-				if(animMoveSpeed < animRunSpeed){
-					animMoveSpeed += animDeltaSpeed * 5;
-				}
-				else{
-					animMoveSpeed -= animDeltaSpeed * 5;
-				}
-				anim.SetFloat("speed", animMoveSpeed);
-			}
-			else{
-				if(animMoveSpeed < animWalkSpeed){
-					animMoveSpeed += animDeltaSpeed;
-				}
-				else{
-					animMoveSpeed -= animDeltaSpeed;
-				}
-				anim.SetFloat("speed", animWalkSpeed);
-			}
-		}
-	}
-
-	void LongJumpingState(){
-        // also since in long jump we dont want to change the character's rotation
-		//float targetRotation = Mathf.Atan2 (input.x, input.y) * Mathf.Rad2Deg + cameraT.eulerAngles.y;
-		//transform.eulerAngles = Vector3.up * Mathf.SmoothDampAngle(transform.eulerAngles.y, targetRotation, ref turnSmoothVelocity, turnSmoothTime);
-
-        // we might not need to recalulate the x and z since universalMovementVector.xz are not changing
-		universalMovementVector.x = initialLongJumpVelocity.x;
-		universalMovementVector.z = initialLongJumpVelocity.y; // z here is actually the y component of this 2D vector
-
-
-        
-        universalMovementVector.y += gravityIncrement;
-		universalMovementVector.y = Mathf.Max(universalMovementVector.y, maxFallVelocity);
-
-		//universalMovementVectorDelta = universalMovementVector * Time.deltaTime;
-		//controller.Move(universalMovementVectorDelta);
-
-		if(universalMovementVector.y <= 0.0f){
-			characterState = CharacterState.FALLING;
-		}
-
-		if(groundPound > 0){
-			characterState = CharacterState.GROUND_POUND;
-		}
-	}
-
-	void HighJumpingState(){
-
-		float targetRotation = Mathf.Atan2 (input.x, input.y) * Mathf.Rad2Deg + cameraT.eulerAngles.y;
-		transform.eulerAngles = Vector3.up * Mathf.SmoothDampAngle(transform.eulerAngles.y, targetRotation, ref turnSmoothVelocity, turnSmoothTime);
-
-		float targetSpeed = ((running || Input.GetKey(KeyCode.E)) ? runSpeed : walkSpeed) * input.magnitude;
-		currentSpeed = Mathf.SmoothDamp (currentSpeed, targetSpeed, ref speedSmoothVelocity, speedSmoothTime);
-
-        universalMovementVector.x = transform.forward.x * currentSpeed;
-        universalMovementVector.z = transform.forward.z * currentSpeed;
-
-
-		universalMovementVector.y += gravityIncrement;
-		universalMovementVector.y = Mathf.Max(universalMovementVector.y, maxFallVelocity);
-
-		//universalMovementVectorDelta = universalMovementVector * Time.deltaTime;
-		//controller.Move(universalMovementVectorDelta);
-
-		if(universalMovementVector.y <= 0.0f){
-			characterState = CharacterState.FALLING;
-		}
-
-		if(groundPound > 0){
-			characterState = CharacterState.GROUND_POUND;
-		}
-	}
-
-	void LowJumpingState(){
-
-		float targetRotation = Mathf.Atan2 (input.x, input.y) * Mathf.Rad2Deg + cameraT.eulerAngles.y;
-		transform.eulerAngles = Vector3.up * Mathf.SmoothDampAngle(transform.eulerAngles.y, targetRotation, ref turnSmoothVelocity, turnSmoothTime);
-
-		float targetSpeed = ((running || Input.GetKey(KeyCode.E)) ? runSpeed : walkSpeed) * input.magnitude;
-		currentSpeed = Mathf.SmoothDamp (currentSpeed, targetSpeed, ref speedSmoothVelocity, speedSmoothTime);
-
-        universalMovementVector.x = transform.forward.x * currentSpeed;
-        universalMovementVector.z = transform.forward.z * currentSpeed;
-
-
-		universalMovementVector.y += gravityIncrement;
-		universalMovementVector.y = Mathf.Max(universalMovementVector.y, maxFallVelocity);
-
-		//universalMovementVectorDelta = universalMovementVector * Time.deltaTime;
-		//controller.Move(universalMovementVectorDelta);
-
-		if(universalMovementVector.y <= 0.0f){
-			characterState = CharacterState.FALLING;
-		}
-
-		if(groundPound > 0){
-			characterState = CharacterState.GROUND_POUND;
-		}
-	}
-
-	void FallingState(){
-		if(isGrounded){
-            //Debug.Log("grounded");
-			characterState = CharacterState.RUNNING; // see if stationary is more appropriate
-		}
-		else{
-
-
-			universalMovementVector.y += gravityIncrement;
-			universalMovementVector.y = Mathf.Max(universalMovementVector.y, maxFallVelocity);
-			//if(fallingMoveBack.magnitude != 0){
-				//universalMovementVector = Vector3.zero;
-			//}
-
-			//controller.Move(universalMovementVector * Time.deltaTime /*+ fallingMoveBack*/);
-			//universalMovementVectorDelta = universalMovementVector * Time.deltaTime + fallingMoveBack;
-			//universalMovementVectorDelta = universalMovementVector * Time.deltaTime + new Vector3(fallingMoveBack.x, 0, fallingMoveBack.z);
-			//controller.Move(universalMovementVectorDelta);
-			//universalMovementVectorDelta = new Vector3(0, fallingMoveBack.y, 0);
-			//controller.Move(universalMovementVectorDelta);
-
-			/*
-			universalMovementVectorDelta = new Vector3(moveBack.x, 0, moveBack.z) * slopeSlipSpeed * Time.deltaTime;
-			controller.Move(universalMovementVectorDelta);
-			
-			universalMovementVectorDelta = new Vector3(0, moveBack.y * 20, 0) * slopeSlipSpeed * Time.deltaTime;
-			controller.Move(universalMovementVectorDelta);
-			*/
-
-
-			//Debug.Log("not grounded");
-			//fallingMoveBack = Vector3.zero;
-			if(groundPound > 0){
-				characterState = CharacterState.GROUND_POUND;
-			}
-		}
-	}
-
-	void SlippingState(){
-		// slipping state will put the character back and down the slope by the moveback vector
-		// it will not accept inputs and will set the state back to stationary, if still on slope,
-		// the collision function will send us back to slipping for a 0 net change
-		// we want all initial movement zeroed out so don't "add" to but rather "set" the universal vector
-		// to avoid issues with collider not allowing movement due to y component forcing character into mesh (maybe)
-		// apply the x and z first and then the y
-		// intially from another state the character moves and the OnControllerColliderHit function is called within after moving
-		// this should then determine a vector to move back and down by and set the state to slipping, on the first slipping call
-		// we move by the moveback vector and thus call the collision function again it will decide wether to swap the state
-
-		characterState = CharacterState.STATIONARY;
-        // the slope has been determined by the raycast and it will have saved it in a vector
-        // we move by this vector, might want to adjust speed by the angle of the slope (60 = 0 and 90 = max)
-		universalMovementVector = slopeVector * slopeSlipSpeed;
-
-
-		//universalMovementVectorDelta = new Vector3(moveBack.x, 0, moveBack.z) * slopeSlipSpeed * Time.deltaTime;
-		//controller.Move(universalMovementVectorDelta);
-
-		//universalMovementVectorDelta = new Vector3(0, moveBack.y * 20, 0) * slopeSlipSpeed * Time.deltaTime;
-		//controller.Move(universalMovementVectorDelta);
-
-		//Debug.Log(moveBack);
-	}
-
-	void GroundPoundState(){
-		// in the ground pound state we halt all motion and start the ground pound animation
-		// a timer is started which will cover the length of the animation plus a little delay
-		// then the rest of the code is allowed to process and the character falls at maximum
-		// velocity. once ground is detected we revert back to stationary
-		if(groundPoundTimer >= groundPoundTimeLimit){
-			// times up proceed with falling
-			if(isGrounded){
-				characterState = CharacterState.STATIONARY;
-				groundPoundTimer = 0.0f;
-			}
-			else{
-                // we are falling during ground pound
-				//controller.Move(new Vector3(0, maxFallVelocity, 0) * Time.deltaTime);
-				universalMovementVector.y = maxFallVelocity;
-				//controller.Move(universalMovementVectorDelta);
-			}
-		}
-		else{
-			groundPoundTimer += Time.deltaTime;
-		}
-	}
-	/*
-	public bool IsGroundPounding(){
-		Debug.Log(characterState);
-		if(characterState == CharacterState.GROUND_POUND){
-			return true;
-		}
-		else{
-			return false;
-		}
-	}
-	*/
-	void LongJump(){
-
-		universalMovementVector = transform.forward * initialLongJumpVelocity.x + new Vector3(0, initialLongJumpVelocity.y, 0);
-
-		//controller.Move(universalMovementVector * Time.deltaTime);
-		//universalMovementVectorDelta = universalMovementVector * Time.deltaTime;
-		//controller.Move(universalMovementVectorDelta);
-
-
-		characterState = CharacterState.JUMPING_LONG;
-	}
-
-	void HighJump(){
-        // keep the x and z as we want to continue to move (but they might be zero anyway since our character has to be stationary)
-        universalMovementVector.y = initialHighJumpVelocity;
-
-		//controller.Move(universalMovementVector * Time.deltaTime);
-		//universalMovementVectorDelta = universalMovementVector * Time.deltaTime;
-		//controller.Move(universalMovementVectorDelta);
-
-		characterState = CharacterState.JUMPING_HIGH;
-	}
-
-	void RegularJump(){
-        // keep the x and z as we want to continue to move
-        universalMovementVector.y = initialJumpVelocity;
-
-		//controller.Move(universalMovementVector * Time.deltaTime);
-		//universalMovementVectorDelta = universalMovementVector * Time.deltaTime;
-		//controller.Move(universalMovementVectorDelta);
-
-		characterState = CharacterState.JUMPING_LOW;
-	}
-    /*
-	bool IsGrounded(){
-		if(Physics.Raycast(transform.position, -Vector3.up, groundDistanceCheck)){
-			//controller.Move(new Vector3(0, -groundDistanceCheck, 0));
-			universalMovementVectorDelta = new Vector3(0, -groundDistanceCheck, 0);
-			controller.Move(universalMovementVectorDelta);
-
-			return true;
-		}
-		else{
-			return false;
-		}
-	}
-    */
-	public enum CharacterState{
-		// if a state switch is detected we finish up the orignal function and swap the state,
-		// and on the NEXT FRAME the new state function will be called (avoid infinite loops)
-
-		// all states can transition to falling
-		// all jumping states transition to falling
-		STATIONARY, // can transition to running, jumping_low, jumping_high
-		RUNNING, // can transition to stationary, jumping_low, jumping_long
-		JUMPING_LOW,
-		JUMPING_LONG,
-		JUMPING_HIGH,
-		FALLING, // upon ground detection will transition to stationary
-		FALLING_NO_CONTROL, // when fall starts without a jump (i.e. bumped off edge)
-		SLIPPING_NO_CONTROL,
-		GROUND_POUND
-		// a function is called to set this value and must clean up a few values before forcing a no input senario
-
-	}
-
-	public void SwitchToNoInput(){
-		// this function is called when we need to deactivate inputs,
-		// must clean up a few things before setting the state to no inputs
-	}
-
-
-	/*
-	Player movement needs to be broken down into states such that one state can transition
-	to another through change in terrain/input. So far states are stationary, running, jumping,
-	falling. Stationary is described as no movement input and can transition into any other
-	state. Running is allowed when a floor is detected and can transition into the other states.
-	Jumping is broken down into what the previous state was, and is subdivided into 3 smaller
-	states. Long jump, regular jump, and high jump. Falling is done either after the initial
-	jumping phase or when a floor is not detected. On each update call the game must first test
-	whether the character is jumping before checking the slope of the floor. This way in impulse
-	to snap the player to the floor is only used if the player is not jumping. Otherwise if the
-	player is jumping then the slope is not checked and an impulse upwards is used instead.
-	During the jumping phase (which is tested each frame by the velocity of the rigid body) a
-	small impulse is added to effectively reduce the effect of gravity during the initial ascent
-	of the jump, gravity is restored to regular value when over the initial jump (and capped off
-	with another value). Another type of impulse is used during the course of a long jump. A 
-	long jump will begin using regular gravity when over the initial height jump of the long
-	jump duration.
-	 */
-
-    
-	void CalculateExtraRotationTranslationVector(){
-		// this function will ray cast downward for a rotating body
-		// if object is tagged ith rotation access its script for
-		// a vector to translate by, the player is grounded when this
-		// function is called
-
-		Vector3 start = transform.position;
-		Vector3 direction = Vector3.down;
-		RaycastHit hit;
-		extraRotationTranslationVector = Vector3.zero;
-		if(Physics.Raycast(start, direction, out hit)){
-			if(hit.transform.tag == "rotatingPlatform"){
-				extraRotationTranslationVector = hit.collider.gameObject.GetComponent<RotatePlatform>().TranslateCharacter(transform.position);
-			}
-		}
-	}
-    /*
-	void OnControllerColliderHit(ControllerColliderHit hit){ // this function in unity is called inside characterController.move()
-		// there are 4 surfaces that can be found here, floor, slippery slope, wall, ceiling
-		// we ray cast in the direction of the hit from position to find the surface hit
-		// 
-
-		if(hit.collider.tag == "rotatingPlatform"){
-			return;
-		}
-		else if(hit.collider.tag == "switch" && characterState == CharacterState.GROUND_POUND){
-			hit.collider.gameObject.GetComponent<buttonCollapse>().Collapse();
-			return;
-		}
-
-
-
-
-		// raycast code to get the normal of the surface
-		RaycastHit ray1;
-		Vector3 direction = hit.point - transform.position;
-		if(Physics.Raycast(transform.position, direction, out ray1, direction.magnitude + forwardDistanceCheck)){
-
-		}
-		else{
-			Debug.Log("error did not find surface");
-		}
-
-
-
-
-		if(ray1.normal.y >= 0.5f){
-			// floor, we do nothing
-		}
-		else if(ray1.normal.y < 0.5f && ray1.normal.y > 0.09f){
-			// slippery slope
-			// gravity wont pull us down we need to move backward first and then down
-			// we might have jumped into the slope and thus are still moving up (in the jumping state)
-			// in which case we ignore the slope, it only applies when grounded or falling
-			if(universalMovementVector.y > 0.0f){
-				return;
-			}
-			else{
-				// we are not falling so now we slide down the slope
-
-				// slippery surface, greater than 60 degrees
-				// we must set the player into the slip state
-				// with a backward/downward vector that is proportional
-				// to the slope of the slipper surface
-
-				// the downward/backward vector is the normal vector with the
-				// value of sqrt(x^2 + z^2) switched with y
-				// so x1 = mx, z1 = mz, and m = y/sqrt(x^2 + z^2)
-				// and y = sqrt(x^2 + z^2)
-				// make y negative to reflect gravity
-				// and move x and z first then y
-				if(characterState != CharacterState.SLIPPING_NO_CONTROL){
-					characterState = CharacterState.SLIPPING_NO_CONTROL;
-					//Debug.Log(hit.normal);
-					// we have not executed a slip
-					float y1 = Mathf.Sqrt(ray1.normal.x * ray1.normal.x + ray1.normal.z * ray1.normal.z);
-					float m = ray1.normal.y/y1;
-					float x1 = ray1.normal.x * m;
-					float z1 = ray1.normal.z * m;
-					// the new vector is ready
-					moveBack = new Vector3(x1, -y1, z1); // this is a unit vector (magnitude = 1)
-				}
-				else{
-					// we are waiting for update to execute our moveback vector, so do nothing
-				}
-
-			}
-
-		}
-		else if(ray1.normal.y < 0.09f && ray1.normal.y > -0.09f){
-			Debug.Log("found a wall" + universalMovementVector.y);
-			// we have found a wall, once again we do nothing if we are in the jumping state
-			// otherwise we must force the player to fall by measuring the slope and creating a moveback
-			// vector and fall down by the amount we lost from the collision which is the current
-			// universalmovementvectorDelta.y - (PrevPos.y - transform.position.y)
-			// apply this directly as is
-			if(universalMovementVector.y > 0.0f){
-				return;
-			}
-			else{
-				// found a wall and not jumping
-
-				fallingMoveBack = transform.position - hit.point;
-				fallingMoveBack.y = 0;
-				Debug.Log(fallingMoveBack);
-
-				/*
-				float y1 = Mathf.Sqrt(ray1.normal.x * ray1.normal.x + ray1.normal.z * ray1.normal.z);
-				float m = ray1.normal.y/y1;
-				float x1 = ray1.normal.x * m;
-				float z1 = ray1.normal.z * m;
-				// the new vector is ready
-
-				fallingMoveBack = new Vector3(-x1, -Mathf.Abs(y1), z1); // this is a unit vector (magnitude = 1)
-				// now multiply it by the value lost
-				// DOUBLE CHECK THIS ITS MAKING Y BECOME POSITIVE
-				Debug.Log("falling move back" + fallingMoveBack + " " + universalMovementVectorDelta.y + " " + PrevPos.y + " " + transform.position.y);
-				fallingMoveBack *= 1.0f;//Mathf.Abs(universalMovementVectorDelta.y - (PrevPos.y - transform.position.y));
-				Debug.Log("falling move back" + fallingMoveBack);
-				
-			}
-		}
-		else{
-			// ceiling do nothing yet, later cancel the momentum and begin falling	
-		}
-
-
-
-
-
-
-
-
-
-
-
-
-
-		/*
-
-		if(hit.collider.tag == "rotatingPlatform"){
-			return;
-		}
-		else if(hit.collider.tag == "switch" && characterState == CharacterState.GROUND_POUND){
-			hit.collider.gameObject.GetComponent<buttonCollapse>().Collapse();
-			return;
-		}
-
-		RaycastHit ray;
-		if(Physics.Raycast(transform.position, Vector3.down, out ray, groundDistanceCheck)){
-			//Debug.Log(ray.normal);
-
-
-			if(ray.normal.y < -0.09f){
-				// ceiling
-			}
-			else if(ray.normal.y < 0.09f){
-				//Debug.Log("wall");
-			}
-			else if(ray.normal.y < 0.5f){
-				// slippery surface, greater than 60 degrees
-				// we must set the player into the slip state
-				// with a backward/downward vector that is proportional
-				// to the slope of the slipper surface
-
-				// the downward/backward vector is the normal vector with the
-				// value of sqrt(x^2 + z^2) switched with y
-				// so x1 = mx, z1 = mz, and m = y/sqrt(x^2 + z^2)
-				// and y = sqrt(x^2 + z^2)
-				// make y negative to reflect gravity
-				// and move x and z first then y
-				if(characterState != CharacterState.SLIPPING_NO_CONTROL){
-					characterState = CharacterState.SLIPPING_NO_CONTROL;
-					//Debug.Log(hit.normal);
-					// we have not executed a slip
-					float y1 = Mathf.Sqrt(ray.normal.x * ray.normal.x + ray.normal.z * ray.normal.z);
-					float m = ray.normal.y/y1;
-					float x1 = ray.normal.x * m;
-					float z1 = ray.normal.z * m;
-					// the new vector is ready
-					moveBack = new Vector3(x1, -y1, z1); // this is a unit vector (magnitude = 1)
-				}
-				else{
-					// we are waiting for update to execute our moveback vector, so do nothing
-				}
-			}
-			else{
-				// regular floor
-			}
-
-
-		}
-		else{
-			//Debug.Log("no floor found");
-			// we did collide with something but it is not a floor, might be stuck on it
-			RaycastHit ray2;
-			Vector3 direction = hit.point - transform.position;
-			if(Physics.Raycast(transform.position, direction, out ray2, direction.magnitude + forwardDistanceCheck)){
-				Debug.Log("wall found");
-				fallingMoveBack = new Vector3(-direction.x, 0, -direction.z) * moveBackValue;
-				Debug.Log(fallingMoveBack);
-			}
-		}
-		
-	}
-    */
-}
-/*
- * {
-    public float speed = 1.0f;
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-
-        // Bit shift the index of the layer (8) to get a bit mask
-        int layerMask = 1 << 8;
-
-        // This would cast rays only against colliders in layer 8.
-        // But instead we want to collide against everything except layer 8. The ~ operator does this, it inverts a bitmask.
-        layerMask = ~layerMask;
-
-        if (Input.GetAxisRaw("Horizontal") > 0)
-        {
-            transform.Translate(transform.forward * speed * Time.deltaTime);
-        }
-        else if(Input.GetAxisRaw("Horizontal") < 0)
-        {
-            transform.Translate(-transform.forward * speed * Time.deltaTime);
-        }
-
-        RaycastHit hit;
-        // Does the ray intersect any objects excluding the player layer
-        if (Physics.Raycast(transform.position, transform.forward, out hit, 1.0f, layerMask))
-        {
-            //Debug.DrawRay(transform.position, transform.forward * hit.distance, Color.yellow);
-            //Debug.Log("Did Hit");
-            transform.Translate(-transform.forward * (1.0f - (hit.point - transform.position).magnitude));
-        }
-        else
-        {
-            Debug.DrawRay(transform.position, transform.forward * 1000, Color.white);
-            Debug.Log("Did not Hit");
-        }
-    }
-}
-*/
